+++ src/Events/Span.php
@@ -85,7 +85,7 @@ class Span extends TraceableEvent implements \JsonSerializable
         if ($this->timer->isNotStarted()) {
             $this->timer->start();
         }
-        $this->timestamp = (int) round($this->timer->getStartTime() * 1000000);
+        $this->timestamp = round($this->timer->getStartTime() * 1000000);
     }
 
     /**
